personal/yxg352/CameronTests/requirements.txt
django-rest-framework
django-rest-swagger